+++ src/main/java/com/hunteryavitz/blockchainapi/services/TransactionService.java
@@ -1,6 +1,7 @@
 package com.hunteryavitz.blockchainapi.services;
 
 import com.hunteryavitz.blockchainapi.entities.Transaction;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 /**
@@ -17,25 +18,21 @@ public class TransactionService {
     /**
      * The blockchainService is an instance of the BlockchainService class.
      */
-    BlockchainService blockchainService;
+    private static BlockchainService blockchainService;
 
     /**
      * The transactionCount is the number of transactions for the purposes of measuring production.
      */
+    @Autowired
     HealthMetricService healthMetricService;
 
     /**
      * The constructor for the TransactionService class.
      */
-    public TransactionService() {
-        this.blockchainService = new BlockchainService();
-        this.healthMetricService = new HealthMetricService();
-    }
-
-    /**
-     * The constructor for the TransactionService class.
-     */
     public void createInitialTransactionPool() {
+        if (blockchainService == null) {
+            blockchainService = new BlockchainService();
+        }
         transactionPool = new Transaction[10];
     }
 
@@ -49,13 +46,11 @@ public class TransactionService {
         if (healthMetricService == null) {
             healthMetricService = new HealthMetricService();
         }
-        healthMetricService.resetBlockCount();
 
         // NOTE: Here to pass the unit tests
         if (blockchainService == null) {
             blockchainService = new BlockchainService();
         }
-        blockchainService.createInitialBlockchain();
 
         for (int i = 0; i < transactionPool.length; i++) {
             if (transactionPool[i] == null) {