+++ src/main/java/com/hunteryavitz/blockchainapi/services/HealthMetricService.java
@@ -1,5 +1,6 @@
 package com.hunteryavitz.blockchainapi.services;
 
+import com.hunteryavitz.blockchainapi.constants.ContaminationLevel;
 import com.hunteryavitz.blockchainapi.entities.healthmetric.BlockchainProduction;
 import com.hunteryavitz.blockchainapi.utils.structures.SlidingWindow;
 import lombok.Getter;
@@ -25,9 +26,15 @@ public class HealthMetricService {
     private static int transactionCount;
 
     /**
+     * The health is an array of integers that represent the number of health metrics of each type.
+     */
+    @Getter
+    private static Integer[] health;
+
+    /**
      * The blockchain production.
      */
-    private final BlockchainProduction blockchainProduction;
+    private static BlockchainProduction blockchainProduction;
 
     /**
      * The constructor.
@@ -36,6 +43,23 @@ public class HealthMetricService {
         blockCount = 0;
         transactionCount = 0;
         blockchainProduction = new BlockchainProduction();
+        health = new Integer[5];
+        for (int i = 0; i < 5; i++) {
+            health[i] = 0;
+        }
+    }
+
+    /**
+     * Creates the health metric service.
+     */
+    public void createHealthMetricService() {
+        blockCount = 0;
+        transactionCount = 0;
+        blockchainProduction = new BlockchainProduction();
+        health = new Integer[5];
+        for (int i = 0; i < 5; i++) {
+            health[i] = 0;
+        }
     }
 
     /**
@@ -81,5 +105,38 @@ public class HealthMetricService {
      */
     public SlidingWindow getProduction() {
         return blockchainProduction.getSlidingWindow();
+    }
+
+    /**
+     * Updates the health.
+     * @param contaminationLevel The contamination level.
+     * @param exception The exception.
+     */
+    public void updateHealth(ContaminationLevel contaminationLevel, Exception exception) {
+        switch (contaminationLevel.toString()) {
+            case "INFO":
+                health[0] = health[0] + 1;
+                System.out.println(health[0]);
+                break;
+            case "WARNING":
+                health[1] = health[1] + 1;
+                System.out.println(health[1]);
+                break;
+            case "ERROR":
+                health[2] = health[2] + 1;
+                System.out.println(health[2]);
+                break;
+            case "CRITICAL":
+                health[3] = health[3] + 1;
+                System.out.println(health[3]);
+                break;
+            case "UNKNOWN":
+                health[4] = health[4] + 1;
+                System.out.println(health[4]);
+                break;
+            default:
+                System.out.println(exception.getMessage());
+                break;
+        }
     }
 }