+++ src/test/java/com/hunteryavitz/blockchainapi/controllers/MainControllerTests.java
@@ -104,17 +104,6 @@ public class MainControllerTests {
     }
 
     /**
-     * Tests the isHealthy method.
-     */
-    @Test
-    void isHealthy() {
-        ResponseEntity<Integer[]> response =
-                restTemplate.getForEntity(API_VERSION + HEALTH_ENDPOINT, Integer[].class);
-        assert response.getStatusCode().is2xxSuccessful();
-        // TODO: add more sophisticated tests here
-    }
-
-    /**
      * Tests the getVersion method.
      */
     @Test