+++ src/main/java/com/hunteryavitz/blockchainapi/constants/NodeStatus.java
@@ -1,0 +1,22 @@
+package com.hunteryavitz.blockchainapi.constants;
+
+/**
+ * The NodeStatus enum represents the status of a node.
+ */
+public enum NodeStatus {
+
+    /**
+     * The node is inactive.
+     */
+    INACTIVE,
+
+    /**
+     * The node is active.
+     */
+    ACTIVE,
+
+    /**
+     * The node is unresponsive.
+     */
+    UNRESPONSIVE
+}