+++ src/test/java/com/hunteryavitz/blockchainapi/controllers/HealthMetricControllerTests.java
@@ -1,5 +1,6 @@
 package com.hunteryavitz.blockchainapi.controllers;
 
+import com.hunteryavitz.blockchainapi.entities.healthmetric.NodeNetworkStatus;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -37,6 +38,11 @@ public class HealthMetricControllerTests {
     private static final String GET_PRODUCTION_HEALTH_ENDPOINT = "/healthMetric/getProductionHealth";
 
     /**
+     * The health metric endpoint.
+     */
+    private static final String GET_NODE_NETWORK_STATUS_ENDPOINT = "/healthMetric/getNodeNetworkStatus";
+
+    /**
      * Tests the updateProduction endpoint.
      */
     @Test
@@ -56,5 +62,19 @@ public class HealthMetricControllerTests {
                 API_VERSION + GET_PRODUCTION_HEALTH_ENDPOINT, String.class);
         assert response.getStatusCode().is2xxSuccessful();
         assert (response.getBody() != null);
+    }
+
+    /**
+     * Tests the getNodeNetworkStatus endpoint.
+     */
+    @Test
+    void testGetNodeNetworkStatus() {
+        ResponseEntity<NodeNetworkStatus> response = restTemplate.getForEntity(
+                API_VERSION + GET_NODE_NETWORK_STATUS_ENDPOINT, NodeNetworkStatus.class);
+        assert response.getStatusCode().is2xxSuccessful();
+        assert (response.getBody() != null);
+
+        NodeNetworkStatus nodeNetworkStatus = response.getBody();
+        assert (nodeNetworkStatus.getNodeNetwork() != null);
     }
 }