+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Blockchain API
 Unreleased
 
-*v0.0.15*
+*v0.0.17*
+
+**11/3/23**
+
+- [ADD Nodes Health Check Service and Endpoint](#54)
 
 **10/31/23**